requirements.txt

torch==2.3.1
torchaudio
numpy==1.26.4
sympy==1.12
transformers==4.41.2
datasets==2.19.1
accelerate==0.30.1
trl==0.8.6
peft==0.11.1

deepspeed==0.14.3
vllm==0.4.3

pyyaml==6.0.1
tqdm==4.66.4
wandb==0.17.0
# A simple library for parsing .env files, useful for managing API keys.
python-dotenv==1.0.1
